Fix REST API
This commit will make changes to the service's API to be more
RESTful. In particular, the response from POSTing to /spam now
sets the location header and returns the posted body with its id.
To allow the user to more easily use the location header, an
alternative way to GET ham has been added through /spam/<id>/ham.

Another addition of this commit is the refactoring of the REST
controllers to keep them separated from domain logic.